Extract overload resolution namespace
Moves overload-resolution types into a dedicated Binding.OverloadResolution namespace and updates call sites/tests. Closes #1365.
