+++ src/Core/CodeAnalysis/Binding/ClrOperatorResolution.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Collections.Generic;
 using System.Reflection;
+using GSharp.Core.CodeAnalysis.Binding.OverloadResolution;
 using GSharp.Core.CodeAnalysis.Symbols;
 using GSharp.Core.CodeAnalysis.Syntax;
 
@@ -15,7 +16,7 @@ namespace GSharp.Core.CodeAnalysis.Binding;
 /// values to CLR operator method names (<c>op_Addition</c>, <c>op_Equality</c>,
 /// <c>op_UnaryNegation</c>, ...) and performs the actual public-static lookup
 /// across the two operand types. The "better function member" tie-break is
-/// delegated to <see cref="OverloadResolution.Resolve{T}"/> so user-defined
+/// delegated to <see cref="ClrOverloadResolution.Resolve{T}"/> so user-defined
 /// operators participate in the same conversion ranking as method calls and
 /// constructors (Stream A).
 /// </summary>
@@ -87,14 +88,14 @@ internal static class ClrOperatorResolution
         }
 
         var argTypes = new[] { leftType?.ClrType, rightType?.ClrType };
-        var outcome = OverloadResolution.Resolve(candidates, argTypes);
-        if (outcome.Outcome == OverloadResolution.ResolutionOutcome.Resolved)
+        var outcome = ClrOverloadResolution.Resolve(candidates, argTypes);
+        if (outcome.Outcome == ResolutionOutcome.Resolved)
         {
             method = outcome.Best;
             return true;
         }
 
-        if (outcome.Outcome == OverloadResolution.ResolutionOutcome.Ambiguous)
+        if (outcome.Outcome == ResolutionOutcome.Ambiguous)
         {
             isAmbiguous = true;
         }
@@ -126,14 +127,14 @@ internal static class ClrOperatorResolution
         }
 
         var argTypes = new[] { operandType.ClrType };
-        var outcome = OverloadResolution.Resolve(candidates, argTypes);
-        if (outcome.Outcome == OverloadResolution.ResolutionOutcome.Resolved)
+        var outcome = ClrOverloadResolution.Resolve(candidates, argTypes);
+        if (outcome.Outcome == ResolutionOutcome.Resolved)
         {
             method = outcome.Best;
             return true;
         }
 
-        if (outcome.Outcome == OverloadResolution.ResolutionOutcome.Ambiguous)
+        if (outcome.Outcome == ResolutionOutcome.Ambiguous)
         {
             isAmbiguous = true;
         }
