+++ src/Core/CodeAnalysis/Binding/PatternBinder.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Collections.Immutable;
 using System.Numerics;
+using GSharp.Core.CodeAnalysis.Binding.OverloadResolution;
 using GSharp.Core.CodeAnalysis.Symbols;
 using GSharp.Core.CodeAnalysis.Syntax;
