+++ test/Core.Tests/CodeAnalysis/Symbols/ReferenceResolverTransitiveClosureTests.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
 using GSharp.Core.CodeAnalysis.Binding;
+using GSharp.Core.CodeAnalysis.Binding.OverloadResolution;
 using GSharp.Core.CodeAnalysis.Symbols;
 using Xunit;
 
@@ -89,8 +90,8 @@ public class ReferenceResolverTransitiveClosureTests : IDisposable
         var candidate = widget.GetMethod("M", BindingFlags.Public | BindingFlags.Static);
         Assert.NotNull(candidate);
 
-        var result = OverloadResolution.Resolve(new[] { candidate }, Array.Empty<Type>());
-        Assert.Equal(OverloadResolution.ResolutionOutcome.NoneApplicable, result.Outcome);
+        var result = ClrOverloadResolution.Resolve(new[] { candidate }, Array.Empty<Type>());
+        Assert.Equal(ResolutionOutcome.NoneApplicable, result.Outcome);
     }
 
     /// <summary>
